Crop the border (stroke) of each <rect> element to the size of the <rect> element itself.
Crop the border (stroke) of each <rect> element to the size
of the <rect> element itself. By default, applying stroke
to an element, draws partially outside of the element borders.
This could cause some confusion for people who look to close
at nearby regions as they could appear to be overlapping while
they are not overlapping each other at all.

Signed-off-by: Gert Hulselmans <gert.hulselmans@kuleuven.be>